Remove old ens impl
