+++ tests/test_ensemble.py
@@ -30,7 +30,7 @@ import typing as t
 import pytest
 
 from smartsim.entity import _mock
-from smartsim.entity._new_ensemble import Ensemble
+from smartsim.entity.ensemble import Ensemble
 from smartsim.entity.strategies import ParamSet
 from smartsim.settings.launchSettings import LaunchSettings
